Create access-comp-data (#132)
* Create access-comp-data

New section that will go into DesignSafe Essentials. This captures the Initial Onboarding and Requesting Allocations that were added after the TAPISv3 transition later in December, and I'm finally getting it into the documentation.

* Update mkdocs.yml

Updating the Nav to include new section in DesignSafe Essentials.

* fix: doc lacks file ext and in wrong dir

* fix: changes from initial review

https://github.com/DesignSafe-CI/DS-User-Guide/pull/132#pullrequestreview-2628191344

* style: remove extra spaces at end of lines

* chore: revert opinionated change

* fix(access-comp-data): broken links - round 1

* fix(access-comp-data): broken links - round 2

* fix(access-comp-data): link only works on prod

* enhance: reduce length of new nav text

but keep page heading as is

---------

Co-authored-by: Wesley B <62723358+wesleyboar@users.noreply.github.com>