Update index.md
I was reviewing yesterday's changes in production this morning, and realized that in the Initial Onboarding section I had left the reference to Interactive VM for OpenSees. However, we have depricated that offering so this edit is to remove that from the text and just have JupyterHub as the example of an app hosted via our virtual machines.